Update patch set 1
Patch Set 1:

Before anyone moans: Yes, it still fails on some repos that we have to push manually as it adds "platform" in the beginning of the path. afaik it is a manifest xml thing, but maybe a mighty Python guy has more time to figure it out than me. 😞

Patch-set: 1