crowdin: switch completely to python3
Cleanup the script and config files for Python3.
Added a crowdin.xml to the platform_manifest with all our translatable repos, which makes syncing easier.

So please, if a repo is added to removed, it must be changed in the platform_manifest's crowdin.xml too. Otherwise the script will fail!
Change-Id: Iac26e195c3850b7c7bfa5c9f32f39e3e5142d116